Separated action buttons from editor header
