--- src/pages/EditorPage/Header/index.tsx
@@ -1,3 +1,0 @@
-import CollaboratorsPopover from './CollaboratorsPopover';
-
-export { CollaboratorsPopover };
